refactor: split server and app
